Rename otter.http to otter.cloud_client
