+++ otter/convergence/service.py
@@ -11,13 +11,13 @@ from pyrsistent import thaw
 
 from twisted.application.service import Service
 
+from otter.cloud_client import TenantScope
 from otter.constants import CONVERGENCE_LOCK_PATH
 from otter.convergence.composition import get_desired_group_state
 from otter.convergence.effecting import steps_to_effect
 from otter.convergence.gathering import get_all_convergence_data
 from otter.convergence.model import ServerState
 from otter.convergence.planning import plan
-from otter.http import TenantScope
 from otter.models.intents import ModifyGroupState
 from otter.util.deferredutils import with_lock
 from otter.util.fp import assoc_obj
