+++ otter/test/convergence/test_service.py
@@ -9,12 +9,12 @@ from twisted.internet.defer import fail
 from twisted.internet.task import Clock
 from twisted.trial.unittest import SynchronousTestCase
 
+from otter.cloud_client import TenantScope, service_request
 from otter.constants import ServiceType
 from otter.convergence.model import (
     CLBDescription, CLBNode, NovaServer, ServerState)
 from otter.convergence.service import (
     Converger, determine_active, execute_convergence)
-from otter.http import TenantScope, service_request
 from otter.models.intents import ModifyGroupState
 from otter.models.interface import GroupState
 from otter.test.convergence.test_planning import server
